backend/requirements.txt
appdirs==1.4.3
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
dj-database-url==0.4.2
Django==1.10.6
django-redis==4.7.0
djangorestframework==3.6.2
envdir==0.7
hiredis==0.2.0
httplib2==0.10.3
incremental==16.10.1
msgpack-python==0.4.8
oauth2client==4.0.0
packaging==16.8
phonenumbers==8.3.2
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
redis==2.10.5
requests==2.13.0
rsa==3.4.2
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
zope.interface==4.3.3
